Create HTML Code
All of the content that will be on the page was created using HTML. The
HTML was broken into 4 main sections plus a footer. An image was linked
in section 1. Four images were linked in section 2. All of the images
were given a width and height of 200px x 200px.