Add css FOR Section-one
First the entire body was made into a flex box. The items were 4 sections
plus a footer. And flex direction was set to column. Section one was made
into a flexbox with a header and a section below it called section-one-two.
The header was made into a flex box. The <ul> in the header was also made
into a flexbox. In the header. A justify-content:space between; was used
to separate these to either side of the page. Section-one-two was made
into a flexbox, with items left and a div with an image. The left side
was made into a flexbox, with flex direction set to column. And once again
justify-content:space-between. was used to push items to either side of
the page. A padding of 50px 200px was given to section one. Img in section
one was changed.